+++ packages/fiori-mcp-server/src/tools/functionalities/generate-fiori-ui-app/command.ts
@@ -2,7 +2,7 @@ import { promises as FSpromises, existsSync } from 'fs';
 import { promisify } from 'util';
 import { exec as execAsync } from 'child_process';
 import { dirname, join } from 'path';
-import type { ExecuteFunctionalitiesInput, ExecuteFunctionalityOutput } from '../../../types';
+import type { ExecuteFunctionalityInput, ExecuteFunctionalityOutput } from '../../../types';
 import { GENERATE_FIORI_UI_APP_ID } from '../../../constant';
 import { findInstalledPackages, type PackageInfo } from '@sap-ux/nodejs-utils';
 import * as z from 'zod';
@@ -58,7 +58,7 @@ const exec = promisify(execAsync);
  * @param params Input parameters for application generation.
  * @returns Application generation execution output.
  */
-export async function command(params: ExecuteFunctionalitiesInput): Promise<ExecuteFunctionalityOutput> {
+export async function command(params: ExecuteFunctionalityInput): Promise<ExecuteFunctionalityOutput> {
     let generatorConfigCAP;
     try {
         generatorConfigCAP = GeneratorConfigSchemaCAP.parse(params.parameters);
@@ -95,7 +95,7 @@ export async function command(params: ExecuteFunctionalitiesInput): Promise<Exec
 
         await FSpromises.mkdir(dirname(outputPath), { recursive: true });
         await FSpromises.writeFile(outputPath, content, { encoding: 'utf8' });
-        const command = `npx -y yo@4 @sap/fiori:headless ${configPath} --force`.trim();
+        const command = `npx -y yo@4 @sap/fiori:headless ${configPath} --force  --skipInstall`.trim();
 
         const { stdout, stderr } = await exec(command, { cwd: targetDir });
         console.log(stdout);
@@ -123,7 +123,7 @@ export async function command(params: ExecuteFunctionalitiesInput): Promise<Exec
     return {
         functionalityId: GENERATE_FIORI_UI_APP_ID,
         status: 'Success',
-        message: 'Generation completed successfully: ' + appPath,
+        message: `Generation completed successfully: ${appPath}. You must run \`npm install\` in ${targetDir} before trying to run the application.`,
         parameters: params.parameters,
         appPath,
         changes: [],