+++ packages/fiori-mcp-server/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/fiori-mcp-server
 
+## 0.0.6
+
+### Patch Changes
+
+-   ac9a2d0: feat: collect telemetry
+
+## 0.0.5
+
+### Patch Changes
+
+-   cbd2a6f: Use Zod schemas to generate input/output JSON schemas and TypeScript types
+
 ## 0.0.4
 
 ### Patch Changes